remove hard coded values
