+++ config.cfg
@@ -36,3 +36,4 @@ UI_AFTER_VERIFY = '/verified/'
 
 CELERY_BROKER_URL = 'amqp://guest@localhost//'
 CELERY_RESULT_BACKEND = 'amqp://guest@localhost//'
+SOCKET_MESSAGE_QUEUE = 'amqp://guest@localhost//'
